+++ packages/natives/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+
+- Added `cache` option to `glob()`, `grep()`, and `fuzzyFind()` to enable shared filesystem scan caching
+- Added `invalidateFsScanCache()` function to manually invalidate filesystem scan cache entries
 
 ## [11.14.0] - 2026-02-12
 ### Added